server/requirements.txt
openai==0.28.1
sqlalchemy==1.4.46
psycopg2-binary==2.9.5
python-dotenv==1.0.0
fastapi==0.95.0
uvicorn==0.21.1
pydantic==1.10.7
jinja2==3.1.2
aiofiles==23.1.0
python-multipart==0.0.6
tiktoken==0.5.1
